rename the pkg name
