+++ pkg/operators/klusterlet/controllers/addonsecretcontroller/controller_test.go
@@ -1,4 +1,4 @@
-package addoncontroller
+package addonsecretcontroller
 
 import (
 	"context"
